backend/requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
coverage==6.4.4
dj-database-url==1.0.0
django==4.1.2
django-environ==0.9.0
django-graphql-jwt==0.3.4
django-jfu==2.0.9
django-model-utils==4.2.0
django-phonenumber-field==6.4.0
faker==14.2.0
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
mailjet-rest==1.3.4
phonenumbers==8.12.54
pillow==10.4.0
promise==2.3
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyjwt==2.4.0
python-dateutil==2.8.2
rx==1.6.1
setuptools==75.1.0
singledispatch==3.7.0
six==1.16.0
slack-sdk==3.18.1
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
untokenize==0.1.1
whitenoise==6.2.0
aniso8601==7.0.0
asgiref==3.5.2
coverage==6.4.4
dj-database-url==1.0.0
django==4.1.2
django-environ==0.9.0
django-graphql-jwt==0.3.4
django-model-utils==4.2.0
django-phonenumber-field==6.4.0
faker==14.2.0
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
mailjet-rest==1.3.4
phonenumbers==8.12.54
pillow==10.4.0
promise==2.3
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyjwt==2.4.0
python-dateutil==2.8.2
rx==1.6.1
setuptools==75.1.0
singledispatch==3.7.0
six==1.16.0
slack-sdk==3.18.1
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
untokenize==0.1.1
whitenoise==6.2.0
